Upgrade seeding system
