+++ starter-code/package.json
@@ -6,7 +6,7 @@
 		"start": "node ./bin/www",
 		"dev": "DEBUG=starter-code:* nodemon ./bin/www",
 		"dev-windows": "nodemon ./bin/www",
-		"seed-db": "node seeds/seedsCelebrities.js"
+		"seed-db": "node seeds/seedDatabase.js"
 	},
 	"dependencies": {
 		"body-parser": "^1.18.3",
