+++ starter-code/seeds/data.js
@@ -1,13 +1,3 @@
-require('dotenv').config();
-
-const mongoose = require('mongoose');
-const Celebrity = require('../models/Celebrity');
-const Movie = require('../models/Movies');
-
-mongoose.connect(process.env.DBURL, { useNewUrlParser: true, useUnifiedTopology: true }).then(x => {
-	console.log(`Connected to database ${x.connections[0].name}`);
-});
-
 const celebrities = [
 	{
 		name: 'Taylor Swift',
@@ -47,22 +37,4 @@ const movies = [
 	}
 ];
 
-async function seedDB(collecName, data) {
-	const count = await collecName.collection.countDocuments({});
-	try {
-		if (count !== 0) {
-			await collecName.collection.drop();
-			console.log('emptied database');
-		}
-		const celebCollection = await collecName.create(data);
-		console.log(`Seed database with ${celebCollection}`);
-	} catch (error) {
-		console.log(`Something went wrong: ${error}`);
-	} finally {
-		mongoose.disconnect();
-		console.log('Disconnected from database');
-	}
-}
-
-seedDB(Celebrity, celebrities);
-seedDB(Movie, movies);
+module.exports = { celebrities, movies };
